Fractline
This generates fractal terrain by starting at a midpoint and iterating recursively along the horizon to adjust heights. I also created trees recursively, by drawing lines which split at the end into bounded-random "branches", with the smallest branches growing "leaves". The sun in the background grows wider as it approaches the horizon, somewhat simulating the lensing of the atmosphere.